# internal/handler/jemaat_handler.go
package handler

import (
	"encoding/json"
	"errors"
	"net/http"
	"strconv"
	"strings"

	"BE_GKMI_NTC_GO/internal/model"
	"BE_GKMI_NTC_GO/internal/repository"
	"BE_GKMI_NTC_GO/internal/response"
	"BE_GKMI_NTC_GO/internal/service"
)

func Jemaat(
	jemaatRepository *repository.JemaatRepository,
	jemaatService JemaatServiceInterface,
) http.HandlerFunc {

	return func(w http.ResponseWriter, r *http.Request) {
		if r.URL.Path == "/api/jemaat/bulk" && r.Method == http.MethodPost {
			BulkJemaat(jemaatService)(w, r)
			return
		}

		switch r.Method {

		case http.MethodGet:

			query := r.URL.Query()

			search := strings.TrimSpace(query.Get("search"))
			jenisKelamin := strings.TrimSpace(query.Get("jenis_kelamin"))
			statusJemaat := strings.TrimSpace(query.Get("status_jemaat"))
			statusDiakonia := strings.TrimSpace(query.Get("status_diakonia"))
			kelompokIbadah := strings.TrimSpace(query.Get("kelompok_ibadah"))

			sortBy := strings.TrimSpace(query.Get("sort_by"))
			sortOrder := strings.TrimSpace(query.Get("sort_order"))

			if sortBy == "" {
				sortBy = "id"
			}

			if sortOrder == "" {
				sortOrder = "asc"
			}

			pageParam := strings.TrimSpace(query.Get("page"))
			limitParam := strings.TrimSpace(query.Get("limit"))

			hasFilter := search != "" ||
				jenisKelamin != "" ||
				statusJemaat != "" ||
				statusDiakonia != "" ||
				kelompokIbadah != "" ||
				query.Get("sort_by") != "" ||
				query.Get("sort_order") != ""

			hasPagination := pageParam != "" || limitParam != ""

			if !hasFilter && !hasPagination {

				jemaatList, err := jemaatService.GetAll(r.Context())
				if err != nil {
					response.Error(
						w,
						http.StatusInternalServerError,
						"Failed to get jemaat",
					)
					return
				}

				_ = response.JSON(
					w,
					http.StatusOK,
					jemaatList,
				)
				return
			}

			page := 1
			limit := 10

			if pageParam != "" {
				value, err := strconv.Atoi(pageParam)
				if err != nil {
					response.Error(
						w,
						http.StatusBadRequest,
						"Invalid page",
					)
					return
				}

				page = value
			}

			if limitParam != "" {
				value, err := strconv.Atoi(limitParam)
				if err != nil {
					response.Error(
						w,
						http.StatusBadRequest,
						"Invalid limit",
					)
					return
				}

				limit = value
			}

			if !hasFilter {
				result, err := jemaatService.GetPaginated(
					r.Context(),
					page,
					limit,
				)

				if err != nil {
					if validationErr, ok := err.(*service.ValidationError); ok {
						response.Error(
							w,
							http.StatusBadRequest,
							validationErr.Message,
						)
						return
					}

					response.Error(
						w,
						http.StatusInternalServerError,
						"Failed to get jemaat",
					)
					return
				}

				_ = response.JSON(
					w,
					http.StatusOK,
					result,
				)
				return
			}

			result, err := jemaatService.GetFilteredPaginated(
				r.Context(),
				search,
				jenisKelamin,
				statusJemaat,
				statusDiakonia,
				kelompokIbadah,
				sortBy,
				sortOrder,
				page,
				limit,
			)

			if err != nil {
				if validationErr, ok := err.(*service.ValidationError); ok {
					response.Error(
						w,
						http.StatusBadRequest,
						validationErr.Message,
					)
					return
				}

				response.Error(
					w,
					http.StatusInternalServerError,
					"Failed to get jemaat",
				)
				return
			}

			_ = response.JSON(
				w,
				http.StatusOK,
				result,
			)

		case http.MethodPost:

			var request model.CreateJemaatRequest

			if err := json.NewDecoder(r.Body).Decode(&request); err != nil {
				response.Error(
					w,
					http.StatusBadRequest,
					"Invalid JSON body",
				)
				return
			}

			if message := request.Validate(); message != "" {
				response.Error(
					w,
					http.StatusBadRequest,
					message,
				)
				return
			}

			jemaat, err := jemaatService.Create(
				r.Context(),
				request,
			)

			if err != nil {
				response.Error(
					w,
					http.StatusInternalServerError,
					"Failed to create jemaat",
				)
				return
			}

			_ = response.JSON(
				w,
				http.StatusCreated,
				jemaat,
			)

		case http.MethodPut:

			idString := strings.TrimPrefix(
				r.URL.Path,
				"/api/jemaat/",
			)

			id, err := strconv.Atoi(idString)
			if err != nil {
				response.Error(
					w,
					http.StatusBadRequest,
					"Invalid jemaat ID",
				)
				return
			}

			var request model.UpdateJemaatRequest

			if err := json.NewDecoder(r.Body).Decode(&request); err != nil {
				response.Error(
					w,
					http.StatusBadRequest,
					"Invalid JSON body",
				)
				return
			}

			if message := request.Validate(); message != "" {
				response.Error(
					w,
					http.StatusBadRequest,
					message,
				)
				return
			}

			jemaat, err := jemaatService.Update(
				r.Context(),
				id,
				request,
			)

			if err != nil {
				if strings.Contains(
					strings.ToLower(err.Error()),
					"not found",
				) {
					response.Error(
						w,
						http.StatusNotFound,
						"Jemaat not found",
					)
					return
				}

				response.Error(
					w,
					http.StatusInternalServerError,
					"Failed to update jemaat",
				)
				return
			}

			_ = response.JSON(
				w,
				http.StatusOK,
				jemaat,
			)

		case http.MethodDelete:

			idString := strings.TrimPrefix(
				r.URL.Path,
				"/api/jemaat/",
			)

			id, err := strconv.Atoi(idString)
			if err != nil {
				response.Error(
					w,
					http.StatusBadRequest,
					"Invalid jemaat ID",
				)
				return
			}

			err = jemaatService.Delete(
				r.Context(),
				id,
			)

			if err != nil {
				if strings.Contains(
					strings.ToLower(err.Error()),
					"not found",
				) {
					response.Error(
						w,
						http.StatusNotFound,
						"Jemaat not found",
					)
					return
				}

				response.Error(
					w,
					http.StatusInternalServerError,
					"Failed to delete jemaat",
				)
				return
			}

			_ = response.JSON(
				w,
				http.StatusOK,
				map[string]string{
					"message": "Jemaat deleted successfully",
				},
			)

		default:

			response.Error(
				w,
				http.StatusMethodNotAllowed,
				"Method not allowed",
			)
		}
	}
}

func BulkJemaat(jemaatService JemaatServiceInterface) http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		if r.Method != http.MethodPost {
			response.Error(w, http.StatusMethodNotAllowed, "Method not allowed")
			return
		}

		var requests []model.CreateJemaatRequest
		if err := json.NewDecoder(r.Body).Decode(&requests); err != nil {
			response.Error(w, http.StatusBadRequest, "Invalid JSON")
			return
		}

		if len(requests) == 0 {
			response.Error(w, http.StatusBadRequest, "Bulk jemaat must not be empty")
			return
		}

		jemaatList := make([]model.Jemaat, 0, len(requests))
		for _, request := range requests {
			jemaatList = append(jemaatList, model.Jemaat{
				NamaPanggilan:  request.NamaPanggilan,
				NamaLengkap:    request.NamaLengkap,
				JenisKelamin:   request.JenisKelamin,
				TanggalLahir:   request.TanggalLahir,
				Domisili:       request.Domisili,
				StatusJemaat:   request.StatusJemaat,
				StatusDiakonia: request.StatusDiakonia,
				KelompokIbadah: request.KelompokIbadah,
			})
		}

		result, err := jemaatService.BulkCreate(r.Context(), jemaatList)
		if err != nil {
			var validationErr *service.ValidationError
			if errors.As(err, &validationErr) {
				response.Error(w, http.StatusBadRequest, validationErr.Message)
				return
			}
			response.Error(w, http.StatusInternalServerError, "Internal server error")
			return
		}

		response.JSON(w, http.StatusCreated, result)
	}
}